Refactor : Ajouter index.ts dans le dossier des pages et des composants
 - Cet index.ts est utilisé pour importer facilement tous les composants que nous voulons ajouter dans un fichier.

Refactorisation : Dashboard.tsx
 - Changement du nom du composant dans le fichier en DashboardPage. (il était Home et causait quelques conflits dans le site)

 Refonte : App.tsx
  - Utilisation du nouveau système d'importation pour être plus facile
