+++ g-sonde/src/App.tsx
@@ -3,35 +3,31 @@ import {
 	IonRouterOutlet,
 	IonSplitPane,
 	setupIonicReact,
-} from "@ionic/react";
-import { IonReactRouter } from "@ionic/react-router";
-import { Redirect, Route } from "react-router-dom";
-import Home from "./pages/Home";
-import About from "./pages/About";
-import DashboardPage from "./pages/User/DashboardPage";
-import Login from "./pages/Login";
-import Aquarium from "./pages/Aquarium";
+} from '@ionic/react';
+import { IonReactRouter } from '@ionic/react-router';
+import { Redirect, Route } from 'react-router-dom';
+import { Home, About, Login, Aquarium, DashboardPage } from './pages';
 
 /* Core CSS required for Ionic components to work properly */
-import "@ionic/react/css/core.css";
+import '@ionic/react/css/core.css';
 
 /* Basic CSS for apps built with Ionic */
-import "@ionic/react/css/normalize.css";
-import "@ionic/react/css/structure.css";
-import "@ionic/react/css/typography.css";
+import '@ionic/react/css/normalize.css';
+import '@ionic/react/css/structure.css';
+import '@ionic/react/css/typography.css';
 
 /* Optional CSS utils that can be commented out */
-import "@ionic/react/css/padding.css";
-import "@ionic/react/css/float-elements.css";
-import "@ionic/react/css/text-alignment.css";
-import "@ionic/react/css/text-transformation.css";
-import "@ionic/react/css/flex-utils.css";
-import "@ionic/react/css/display.css";
+import '@ionic/react/css/padding.css';
+import '@ionic/react/css/float-elements.css';
+import '@ionic/react/css/text-alignment.css';
+import '@ionic/react/css/text-transformation.css';
+import '@ionic/react/css/flex-utils.css';
+import '@ionic/react/css/display.css';
 
 /* Theme variables */
-import "./theme/variables.css";
+import './theme/variables.css';
 
-import "../styles/tailwind.css";
+import '../styles/tailwind.css';
 
 setupIonicReact();
 
@@ -40,19 +36,19 @@ const App: React.FC = () => {
 		<IonApp>
 			<IonReactRouter>
 				<IonRouterOutlet>
-					<Route exact path={"/"}>
+					<Route exact path={'/'}>
 						<Home></Home>
 					</Route>
-					<Route path={"/user/home"}>
+					<Route path={'/user/home'}>
 						<DashboardPage></DashboardPage>
 					</Route>
-					<Route path={"/aquarium"}>
+					<Route path={'/aquarium'}>
 						<Aquarium></Aquarium>
 					</Route>
-					<Route path={"/login"}>
+					<Route path={'/login'}>
 						<Login></Login>
 					</Route>
-					<Route path={"/about"}>
+					<Route path={'/about'}>
 						<About></About>
 					</Route>
 				</IonRouterOutlet>
