+++ g-sonde/src/pages/User/DashboardPage.tsx
@@ -11,54 +11,54 @@ import {
 	IonTitle,
 	IonToolbar,
 	isPlatform,
-} from "@ionic/react";
-import axios from "axios";
-import React from "react";
+} from '@ionic/react';
+import axios from 'axios';
+import React from 'react';
 
-const Home: React.FC = () => {
+const DashboardPage: React.FC = () => {
 	async function logout() {
-		window.location.href = "/";
+		window.location.href = '/';
 	}
 	const [data, setData] = React.useState();
 
 	async function getData() {
 		axios
-			.get("http://localhost:3001/api/user", {
+			.get('http://localhost:3001/api/user', {
 				withCredentials: true,
 			})
 			.then((res) => {
 				setData(res.data.message);
 			})
 			.catch((err) => {
-				window.location.href = "/login";
+				window.location.href = '/login';
 			});
 	}
 
 	return (
 		<>
-			<IonMenu contentId="main-content">
+			<IonMenu contentId='main-content'>
 				<IonHeader>
 					<IonToolbar>
 						<IonTitle>Menu Content</IonTitle>
 					</IonToolbar>
 				</IonHeader>
-				<IonContent className="ion-padding">
+				<IonContent className='ion-padding'>
 					This is the menu content.
-					<IonButton expand="block" onClick={logout}>
+					<IonButton expand='block' onClick={logout}>
 						Logout
 					</IonButton>
 				</IonContent>
 			</IonMenu>
-			<IonPage id="main-content">
+			<IonPage id='main-content'>
 				<IonHeader>
 					<IonToolbar>
-						<IonButtons slot="start">
+						<IonButtons slot='start'>
 							<IonMenuButton></IonMenuButton>
 						</IonButtons>
 						<IonTitle>Menu</IonTitle>
 					</IonToolbar>
 				</IonHeader>
-				<IonContent className="ion-padding">
+				<IonContent className='ion-padding'>
 					Tap the button in the toolbar to open the menu.
 					<br />
 					<IonButton onClick={getData}>Get Data</IonButton>
@@ -69,4 +69,4 @@ const Home: React.FC = () => {
 	);
 };
 
-export default Home;
+export default DashboardPage;
