+++ g-sonde/src/pages/index.ts
@@ -1,0 +1,5 @@
+export { default as Home } from './Home';
+export { default as About } from './About';
+export { default as Aquarium } from './Aquarium';
+export { default as Login } from './Login';
+export { default as DashboardPage } from './User/DashboardPage';
